refactor: #15 - Component index 경로 설정
