+++ src/components/index.ts
@@ -1,0 +1,7 @@
+export { default as DatePickerComponent } from './DatePickerComponent';
+export { default as HeaderBar } from './HeaderBar';
+export { default as Message } from './Message';
+export { default as NavBar } from './NavBar';
+export { default as Slider } from './Slider';
+export { default as Spinner } from './Spinner';
+export { default as TabBar } from './TabBar';
